+++ src/lib.rs
@@ -6,6 +6,7 @@ pub(crate) mod kubectl;
 pub(crate) mod process;
 pub(crate) mod tailscale;
 pub(crate) mod time_ext;
+pub(crate) mod user_host;
 
 pub mod cli;